[Refactor]: 드롭다운 그룹 조회 응답 DTO 패키지 변경
