+++ src/main/java/scs/planus/domain/group/service/GroupService.java
@@ -5,12 +5,13 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import scs.planus.domain.group.dto.GroupBelongInResponseDto;
 import scs.planus.domain.group.dto.GroupCreateRequestDto;
+import scs.planus.domain.group.dto.GroupDetailUpdateRequestDto;
 import scs.planus.domain.group.dto.GroupGetResponseDto;
+import scs.planus.domain.group.dto.GroupNoticeUpdateRequestDto;
 import scs.planus.domain.group.dto.GroupResponseDto;
 import scs.planus.domain.group.dto.GroupTagResponseDto;
-import scs.planus.domain.group.dto.*;
+import scs.planus.domain.group.dto.mygroup.GroupBelongInResponseDto;
 import scs.planus.domain.group.entity.Group;
 import scs.planus.domain.group.entity.GroupMember;
 import scs.planus.domain.group.entity.GroupTag;
@@ -169,18 +170,6 @@ public class GroupService {
         groupMemberRepository.save(groupMember);
 
         return GroupResponseDto.of( group );
-    }
-
-    public List<GroupBelongInResponseDto> getMyGroupsInDropDown(Long memberId) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new PlanusException(NONE_USER));
-
-        List<GroupMember> groupMembers = groupMemberRepository.findAllByActiveGroupAndMemberId(member.getId());
-        List<GroupBelongInResponseDto> responseDtos = groupMembers.stream()
-                .map(groupMember -> GroupBelongInResponseDto.of(groupMember.getGroup()))
-                .collect(Collectors.toList());
-
-        return responseDtos;
     }
 
     private void validateLeaderPermission( Member member, Group group ) {
